Revert "fixed rate int bug"
This reverts commit 819f9fe5701336866477cfd0b5847168df278995.

+++ src/main/java/com/example/CurrencyExchange/model/Transaction.java
@@ -183,11 +183,10 @@ public class Transaction {
 
             ObjectMapper mapper = new ObjectMapper();
             JsonNode rateNode = mapper.readTree(response.toString());
-            this.rate = (float) rateNode.get(currencyCodeTo).asDouble();
-            this.price = rate * amount;
+            this.rate = rateNode.get(currencyCodeTo).asLong();
+            this.price = rate*amount;
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
     }
-
 }
